clr: Refactor relationship between Interpreter and Frame.
A Frame should not be an Interpreter. Instead, there should be only one
Interpreter per set of Frames (so we don't have to copy every user-defined
function every time we make a new frame), and it should be immutable after
the first frame is committed (or later frames could alter state in a
way that is visible to earlier frames, preventing rewind+play from working
properly).

We use a new Context type to track the current frame. This type is extensible
and may in the future be used for other information.